crud changes with access to auther only

+++ src/Controller/JobController.php
@@ -4,7 +4,6 @@ namespace App\Controller;
 
 use App\Entity\Job;
 use App\Form\JobType;
-use App\Repository\JobRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,27 +12,22 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/job')]
 class JobController extends AbstractController
 {
-    #[Route('/', name: 'job_index', methods: ['GET'])]
-    public function index(JobRepository $jobRepository): Response
-    {
-        return $this->render('job/index.html.twig', [
-            'jobs' => $jobRepository->findAll(),
-        ]);
-    }
-
-    #[Route('/new', name: 'job_new', methods: ['GET','POST'])]
+    #[Route('/new', name: 'job_new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_RECRUTER');
+
         $job = new Job();
         $form = $this->createForm(JobType::class, $job);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            // $job->setRecruter($this->getUser());
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($job);
             $entityManager->flush();
 
-            return $this->redirectToRoute('job_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('my-jobs', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('job/new.html.twig', [
@@ -50,16 +44,20 @@ class JobController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'job_edit', methods: ['GET','POST'])]
+    #[Route('/{id}/edit', name: 'job_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Job $job): Response
     {
+        if ($this->getUser()->getId() !== $job->getRecruter()->getId() || in_array('ROLE_RECRUTER', $this->getUser()->getRoles())) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(JobType::class, $job);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('job_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('my-jobs', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('job/edit.html.twig', [
@@ -71,12 +69,16 @@ class JobController extends AbstractController
     #[Route('/{id}', name: 'job_delete', methods: ['POST'])]
     public function delete(Request $request, Job $job): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$job->getId(), $request->request->get('_token'))) {
+        if ($this->getUser()->getId() !== $job->getRecruter()->getId() || in_array('ROLE_RECRUTER', $this->getUser()->getRoles())) {
+            throw $this->createAccessDeniedException();
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $job->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($job);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('job_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('my-jobs', [], Response::HTTP_SEE_OTHER);
     }
 }
